Add BaseTaskObjC to podspec source files
In order to ensure that BaseTaskObjC is included when this library is
built using cocoapods, this commit adds it to the podspec. In addition,
it modifies a few properties on objects to ensure they are publicly
accessible instead of marked as internal. And finally, it updates a few
build settings to ensure it can be included in other libraries and
applications correctly.